products/api_endpoints/Category/CategoryDelete/serializers.py
from rest_framework import serializers

from products.models import Category


class CategoryDeleteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ["id"]
